clean up network listeners
