+++ app/store/middleware/listener.ts
@@ -1,8 +1,5 @@
 import { createListenerMiddleware, TypedStartListening } from '@reduxjs/toolkit'
-import { setNetwork } from '@avalabs/wallet-react-components'
-import NetworkService from 'services/network/NetworkService'
-import { onStorageReady } from '../actions'
-import { setActive, selectActiveNetwork } from '../network'
+import { addNetworkListeners } from '../network'
 import type { RootState, AppDispatch } from '../index'
 
 export type AppStartListening = TypedStartListening<RootState, AppDispatch>
@@ -11,32 +8,6 @@ const listener = createListenerMiddleware()
 
 const startListening = listener.startListening as AppStartListening
 
-// APP LIFECYCLE LISTENERS
-startListening({
-  actionCreator: onStorageReady,
-  effect: async (action, listenerApi) => {
-    const state = listenerApi.getState()
-    NetworkService.onStorageReady(state)
-
-    // wallet-react-components sets MAINNET as the active network on app start
-    // we need to set it back to whatever network persisted in our app
-    // TODO: remove this once network refactor is done
-    const network = selectActiveNetwork(state)
-    setNetwork(network)
-  }
-})
-
-// NETWORK LISTENERS
-startListening({
-  actionCreator: setActive,
-  effect: async (action, listenerApi) => {
-    const state = listenerApi.getState()
-
-    // TODO: remove this once network refactor is done
-    // for now, still need to also set active network in wallet-react-components
-    const network = selectActiveNetwork(state)
-    setNetwork(network)
-  }
-})
+addNetworkListeners(startListening)
 
 export { listener }
